refactor: move ChannelState into its own module
