+++ ably/realtime/realtime_channel.py
@@ -1,11 +1,10 @@
 import asyncio
-from dataclasses import dataclass
 import logging
-from typing import Optional
 
 from ably.realtime.connection import ConnectionState
 from ably.transport.websockettransport import ProtocolMessageAction
 from ably.rest.channel import Channel
+from ably.types.channelstate import ChannelState, ChannelStateChange
 from ably.types.message import Message
 from ably.util.eventemitter import EventEmitter
 from ably.util.exceptions import AblyException
@@ -14,16 +13,6 @@ from enum import Enum
 from ably.util.helper import Timer, is_callable_or_coroutine
 
 log = logging.getLogger(__name__)
-
-
-class ChannelState(str, Enum):
-    INITIALIZED = 'initialized'
-    ATTACHING = 'attaching'
-    ATTACHED = 'attached'
-    DETACHING = 'detaching'
-    DETACHED = 'detached'
-    SUSPENDED = 'suspended'
-    FAILED = 'failed'
 
 
 class Flag(int, Enum):
@@ -42,14 +31,6 @@ class Flag(int, Enum):
 
 def has_flag(message_flags: int, flag: Flag):
     return message_flags & flag > 0
-
-
-@dataclass
-class ChannelStateChange:
-    previous: ChannelState
-    current: ChannelState
-    resumed: bool
-    reason: Optional[AblyException] = None
 
 
 class RealtimeChannel(EventEmitter, Channel):
@@ -328,7 +309,7 @@ class RealtimeChannel(EventEmitter, Channel):
             flags = msg.get('flags')
             error = msg.get("error")
             exception = None
-            resumed = None
+            resumed = False
 
             if error:
                 exception = AblyException(error.get('message'), error.get('statusCode'), error.get('code'))
